+++ src/application/Main.java
@@ -1,0 +1,5 @@
+package application;
+
+public class Main {
+
+}
